1. ✅ 修复缺失模块 - 在 lib.rs 中添加了 error 和 import 模块声明   2. ✅ 修复格式字符串错误 - 在 es_client.rs:534 添加了缺失的格式说明符 {}   3. ✅ 修复类型转换问题 - 在 commands.rs 中将 u64 转换为 i32 (添加 as i32)   4. ✅ 修复 Value::from 字符串转换 - 将 Value::from(string) 改为 Value::from(string.as_str())   5. ✅ 修复临时值借用问题 - 为临时 Map 创建了绑定变量   6. ✅ 移除未使用的导入 - 删除了 export.rs 中的 HashMap 和 Worksheet 导入

+++ src-tauri/src/commands.rs
@@ -398,8 +398,8 @@ pub async fn export_search_results(
         let current_size = std::cmp::min(batch_size, remaining);
 
         let mut search_query = request.query.clone();
-        search_query.from = Some(current_from);
-        search_query.size = Some(current_size);
+        search_query.from = Some(current_from as i32);
+        search_query.size = Some(current_size as i32);
 
         match client.search(search_query).await {
             Ok(result) => {
